feat: RemovePostService should throw an error if post with provided id does not exist

+++ src/modules/post/services/RemovePostService.ts
@@ -1,3 +1,4 @@
+import APIError from '@src/errors/APIError';
 import { IPostsRepository } from '../repositories/interface/IPostsRepository';
 import { IRemovePostService } from './interfaces/IRemovePostService';
 
@@ -9,7 +10,9 @@ export class RemovePostService implements IRemovePostService {
   }
 
   async execute(id: string): Promise<void> {
-    await this.#postsRepository.getById(id);
+    const post = await this.#postsRepository.getById(id);
+
+    if (!post) throw new APIError(404, 'Post with provided id does not exist.');
 
     await this.#postsRepository.delete(id);
   }
